refactor(plop): move angular command file to plop template
