+++ plop-templates/angular-component.js
@@ -1,0 +1,41 @@
+module.exports = (cwd) => ({
+  prompts: [
+    {
+      type: 'input',
+      name: 'name',
+      message: 'component name please',
+    },
+    {
+      type: 'input',
+      name: 'module',
+      message: 'module name please',
+      default: `${getCurrentPortainerModule(cwd)}`,
+      // when: false
+    },
+  ], // array of inquirer prompts
+  actions: [
+    {
+      type: 'add',
+      path: `{{cwd}}/{{dashCase name}}/index.js`,
+      templateFile: './plop-templates/component.js.hbs',
+    },
+    {
+      type: 'add',
+      path: `{{cwd}}/{{dashCase name}}/{{dashCase name}}.controller.js`,
+      templateFile: './plop-templates/component-controller.js.hbs',
+    },
+    {
+      type: 'add',
+      path: `{{cwd}}/{{dashCase name}}/{{dashCase name}}.html`,
+      templateFile: './plop-templates/component.html.hbs',
+    },
+  ], // array of actions
+});
+
+function getCurrentPortainerModule(cwd) {
+  const match = cwd.match(/\/app\/([^\/]*)(\/.*)?$/);
+  if (!match || !match.length || match[1] === 'portainer') {
+    return 'app';
+  }
+  return match[1];
+}
